Allow to build without vendored openssl, allow to build syntect with regex-onig (#1323)
* allow to build syntect with regex-onig

Syntect supports two regex engines:

* regex-fancy: a pure-rust regex engine based on the fancy-regex
* regex-onig: a regex engine based on the oniguruma C library

From the syntect's Readme:

> The advantage of fancy-regex is that it does not require the onig
> crate which requires building and linking the Oniguruma C library.
> Many users experience difficulty building the onig crate, especially
> on Windows and Webassembly.

> As far as our tests can tell this new engine is just as correct, but
> it hasn't been tested as extensively in production. It also currently
> seems to be about half the speed of the default Oniguruma engine

Oniguruma engine is faster than the fancy-regex engine and the syntect
project chose the latter as the default only to avoid difficulties with
linking Oniguruma (C library) on some platforms. This is not an issue
for linux distributions - linking against system-provided shared
library is preferred to bundled libraries.

Moreover, gitui built with Oniguruma instead of fancy-regex is by 25%
smaller.

This commit adds two cargo features, regex-fancy and regex-onig, to
enable respective syntect features. The former is enabled by default.

* allow to build without vendored openssl

Vendoring (bundling) openssl library is very bad for security and
Linux distributions forbid it. The aim of this change is to simplify
packaging gitui in linux distros.

Co-authored-by: extrawurst <776816+extrawurst@users.noreply.github.com>